backend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyPDF2==3.0.1
Pillow==9.5.0
python-dotenv==0.19.0
flask-cors==4.0.0
pdf2image==1.16.3
numpy==1.24.0